Update README.md
Let’s focus the main Readme.md on things that are not polemical and on what unite us all.
Let’s use a separate contribute.md or guidelines.md to convey the philosophy and what to do / not to do. Also probably better to wait for something to happen before coming up with rules/recommendations.